Fix Bug where the Premium Button is not showing when the user buys the premium
